pyproject.toml
[tool.poetry]
name            = "citation-network-gnn"
version         = "0.1.0"
description     = "Graph neural network approach to predicting missing citations in academic networks"
authors         = ["Barbara Hidalgo <you@example.com>"]
license         = "MIT"
readme          = "README.md"
packages        = [
  { include = "src" }
]
package-mode = false

[tool.poetry.dependencies]
python          = ">=3.10,<3.13"

# ── Core ML & Graph Libraries ──────────────────────────────
torch           = "^2.0"           # PyTorch for TransE embeddings
networkx        = "^3.2"           # Graph construction and analysis
pandas          = "^2.2"           # Data manipulation
numpy           = "^1.26"          # Numerical computing
scikit-learn    = "^1.3"           # ML evaluation metrics
neo4j           = "^5.19"          # Database driver

# ── Visualization & Analysis ───────────────────────────────
matplotlib      = "^3.7"           # Static plotting
seaborn         = "^0.12"          # Statistical visualization
jupyter         = "^1.0"           # Notebook environment

# ── Utilities ──────────────────────────────────────────────
python-dotenv   = "^1.0"           # Environment management
tqdm            = "^4.65"          # Progress bars

# ── Development / QA only ──────────────────────────────────
[tool.poetry.group.dev.dependencies]
black           = "^24.4"
isort           = "^5.13"
flake8          = "^7.0"
mypy            = "^1.10"
pytest          = "^8.1"
pytest-mock     = "^3.14"
pre-commit      = "^3.7"

# ── Build backend ──────────────────────────────────────────
[build-system]
requires        = ["poetry-core>=1.9.0"]
build-backend   = "poetry.core.masonry.api"

# ── Tooling configuration hints (optional) ─────────────────
[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version      = "3.10"
strict              = true
ignore_missing_imports = true

